Add files via upload
This is a star sign generator. It uses the users input(birth date and month) to determine their star sign. It then displays the users name(taken from an input)  along with their star sign and meaning of it. It does this by accessing a list of menaning from a dictionary.