Add files via upload
This is a rock, paper, scissors game. The code uses the import random module random.choice() to choose a random choice between rock, paper, scissors as the computers move.
It then stores the computer move as a variable and the user chooses between rock, paper, scissors (input). This is then compared with the computer move using if statements and the final ouput will either be a win, a lose or a draw.