Add files via upload
This code lets the user create their own password. the password has to be 10 characters or more or the the code will ask the user to reenter the password again. This is done through a while loop. The user than enters their name, surname and birth year and the code uses this information to generate a username for you.
The final output will be the username created by the code and the password you created.